+++ CHANGELOG.md
@@ -17,7 +17,8 @@ list[?10] == null;
 
 "hello"[?20] == null;
 ```
-- In REPL, user input is syntax highlighted (https://github.com/buzz-language/buzz/issues/217)
+- User input is syntax highlighted in REPL (https://github.com/buzz-language/buzz/issues/217)
+- REPL handles multilines input (https://github.com/buzz-language/buzz/issues/218)
 
 ## Modified
 - Enum can now have `rg`, `ud`, `void`, `pat` has value type